# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/datasets": {
      "get": {
        "operationId": "listDatasets",
        "parameters": [],
        "tags": ["datasets"],
        "description": "List all available datasets",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "uploadDataset",
        "parameters": [],
        "tags": ["datasets"],
        "description": "Upload a dataset file (JSON, JSONL, or CSV)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [true],
                      "type": "boolean"
                    }
                  },
                  "required": ["success"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/datasets/import-hf": {
      "post": {
        "operationId": "importHfDataset",
        "parameters": [],
        "tags": ["datasets"],
        "description": "Import a dataset from HuggingFace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["config", "dataset", "split"],
                "type": "object",
                "properties": {
                  "config": {
                    "type": "string",
                    "default": "default"
                  },
                  "dataset": {
                    "type": "string",
                    "minLength": 1
                  },
                  "split": {
                    "type": "string",
                    "default": "train"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "filename": {
                      "type": "string"
                    },
                    "success": {
                      "enum": [true],
                      "type": "boolean"
                    }
                  },
                  "required": ["filename", "success"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetEntry": {
        "type": "object",
        "properties": {
          "extension": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          }
        },
        "required": ["extension", "name", "sizeBytes"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use Authorization: Bearer <token> for session tokens or API keys."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Use X-API-Key: <key>."
      }
    }
  }
}